% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{AR.reml}
\alias{AR.reml}
\title{Perform AR(1) REML on a model with covariates u}
\usage{
AR.reml(formula, data = list(), par, x, u)
}
\arguments{
\item{u}{}
}
\value{

}
\description{
This function performs AR(1) REML using ... [Description]
last modified on 01-April-2020
}
