% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{AR_df}
\alias{AR_df}
\title{Build AR data frame}
\usage{
AR_df(x, t, Z = NULL)
}
\arguments{
\item{x}{length p time series vector}

\item{t}{length p vector of time}

\item{Z}{optional vector or design matrix of covariates. if only one
covariate, \code{length(Z)} = p. If more than one, \code{nrow(Z)} = p}
}
\value{

}
\description{
Build AR data frame
}
