% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_reml.R
\name{AR_funct}
\alias{AR_funct}
\title{restricted maximum likelihood of an AR model}
\usage{
AR_funct(par, x, U, LL.only = TRUE)
}
\arguments{
\item{par}{AR parameter value}

\item{x}{vector of time series (response)}

\item{U}{model matrix (predictors)}
}
\value{
if \code{LL.only = TRUE}: returns log-likelihood of \code{par}
given \code{x} and \code{u}.

otherwise, returns a "remoteAR.pixel" object which is a list of a
regression coefficeint table (\code{$coef}), the REML AR parameter (\code{$b}),
model MSE (\code{$MSE}),
estimated coefficeint covariance matrix (\code{$s2beta}),
and the model log-likelihood (\code{$logLik}).
}
\description{
restricted maximum likelihood of an AR model
}
\details{
used by \code{fitAR()}
}
\examples{
time = 1:30
x = rnorm(31)
x = x[2:31] + x[1:30] + 0.3*time #AR(1) process + time trend
u = model.matrix(formula(x ~ time))
AR_funct(par = .2, x, u, LL.only = TRUE)
AR_funct(par = .2, x, u, LL.only = FALSE)
}
