% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{as.data.frame.remoteAR}
\alias{as.data.frame.remoteAR}
\title{return dataframe frome remoteAR.map object}
\usage{
\method{as.data.frame}{remoteAR}(obj)
}
\arguments{
\item{obj}{remoteAR object}
}
\value{

}
\description{
return dataframe frome remoteAR.map object
}
