% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{as.data.frame.remoteCLS}
\alias{as.data.frame.remoteCLS}
\title{return remoteCLS obj as data frame}
\usage{
\method{as.data.frame}{remoteCLS}(obj)
}
\arguments{
\item{obj}{remoteCLS map object (see \code{fitCLS.map()})}
}
\value{
as.data.frame(obj$time.coef)
}
\description{
return remoteCLS obj as data frame
}
