% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_mod.R
\name{build_mod}
\alias{build_mod}
\title{Build a usable model matrix from a formula and data}
\usage{
build_mod(formula, data, subset, na.action, contrasts = NULL)
}
\arguments{
\item{formula}{formula to use}

\item{data}{data to use}

\item{subset}{subset of the data}

\item{na.action}{what to do with NAs}

\item{contrasts}{optional linear contrasts}
}
\value{
a remoteGLS object
}
\description{
Build a usable model matrix from a formula and data
}
