% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_partition.R
\name{calc_df}
\alias{calc_df}
\title{calculate degrees of freedom for partitioned GLS}
\usage{
calc_df(part.size, p, p0)
}
\arguments{
\item{part.size}{number of pixels in each partition}

\item{p}{number of predictors in alternate model}

\item{p0}{number of parameters in null model}
}
\description{
calculate degrees of freedom for partitioned GLS
}
