% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_partition.R
\name{calc_dfpart}
\alias{calc_dfpart}
\title{calculate degrees of freedom for partitioned GLS}
\usage{
calc_dfpart(partsize, p, p0)
}
\arguments{
\item{p}{number of predictors in alternate model}

\item{p0}{number of parameters in null model}

\item{part.size}{number of pixels in each partition}
}
\description{
calculate degrees of freedom for partitioned GLS
}
\examples{
calc_df(partsize = 2000, p = 4, p0 = 1)
}
