% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLS_methods.R
\name{chisqr.partGLS}
\alias{chisqr.partGLS}
\title{Conduct a chisqr test of "partGLS" object}
\usage{
\method{chisqr}{partGLS}(x, ...)
}
\arguments{
\item{x}{"remoteGLS" object}

\item{...}{additional arguments passed to print}
}
\value{
a p-value for the correlated chisqr test
}
\description{
Conduct a correlated chi-square test on a partitioned GLS
}
