% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLS_methods.R
\name{chisqr}
\alias{chisqr}
\title{Conduct a chi-squared test}
\usage{
chisqr(x, ...)
}
\arguments{
\item{x}{object on which to conduct the test}

\item{...}{additional arguments}
}
\description{
generic S3 method for a chi-squared test
}
