% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{CLS.fit}
\alias{CLS.fit}
\title{Fit a CLS model to a matrix of time series}
\usage{
CLS.fit(X, t.scale)
}
\arguments{
\item{t.scale}{}
}
\value{

}
\description{
CLS.fit fits a CLS model using lm() to a matrix of time series located in
each row
}
