% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{coef.remoteAR}
\alias{coef.remoteAR}
\title{Extract coefficients from remoteAR object}
\usage{
\method{coef}{remoteAR}(obj)
}
\arguments{
\item{obj}{remoteAR object}
}
\value{
coefficeint table
}
\description{
Extract coefficients from remoteAR object
}
