% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{coef.remoteCLS}
\alias{coef.remoteCLS}
\title{Extract coefficeints from remoteCLS object}
\usage{
\method{coef}{remoteCLS}(obj)
}
\arguments{
\item{obj}{remoteCLS object}
}
\value{
coefficient data frame
}
\description{
Extract coefficeints from remoteCLS object
}
