% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cor_chisq}
\alias{cor_chisq}
\title{Correlated chi-squared test}
\usage{
cor_chisq(Fmean, rSSR, df1, npart)
}
\arguments{
\item{Fmean}{average partition F value}

\item{rSSR}{cross-partition regression sum of squares}

\item{df1}{first degrees of freedom}

\item{npart}{number of partitions used to split the data}
}
\description{
Correlated chi-squared test
}
\details{
performs a correlated chi-squared test on cross-partition GLS
results.
}
