% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{cor_chisq.test}
\alias{cor_chisq.test}
\title{correlated chi-squared test of partitioned GLS}
\usage{
cor_chisq.test(obj)
}
\arguments{
\item{obj}{remoteGLS.parts object}
}
\value{
p value
}
\description{
correlated chi-squared test of partitioned GLS
}
