% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{cor_F.test}
\alias{cor_F.test}
\title{correlated F test of partitioned GLS}
\usage{
cor_F.test(obj, nboot = 1000)
}
\arguments{
\item{obj}{remoteGLS.parts object}

\item{nboot}{number of bootstrap iterations}
}
\value{
P value
}
\description{
correlated F test of partitioned GLS
}
