% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{cor_t}
\alias{cor_t}
\title{analytical t-test for partitioned GLS}
\usage{
cor_t(coefs, part.SEs, rcoef, df2, npart)
}
\arguments{
\item{coefs}{the coefficients to test - averaged across partitions}

\item{part.SEs}{matrix of partition standard errors (columns are partitions)}

\item{rcoef}{cross-coefficient averaged across partition pairs}

\item{df2}{second degrees of freedom from partitioned GLS}

\item{npart}{number of partitions the data was split into}
}
\description{
analytical t-test for partitioned GLS
}
