% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{correlated.F.bootstrap}
\alias{correlated.F.bootstrap}
\title{bootstrap multiple-comparison corrected p-values}
\usage{
correlated.F.bootstrap(Fmean.obs, rSSR, rSSE, df1, df2, npart, nboot = 2000)
}
\arguments{
\item{nboot}{}
}
\value{

}
\description{
correlated.F.bootstrap gives the multiple-comparisons-corrected p-values
and also performs a bootstrap to get the distribution of a correlated F test.
It might be possible to work this out algebraically.
}
