% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worker-functions_GLS.R
\name{crosspart_worker}
\alias{crosspart_worker}
\alias{crosspart_worker_R}
\title{Worker function 2 for partitioned GLS}
\usage{
crosspart_worker(
  xxi,
  xxj,
  xxi0,
  xxj0,
  invChol_i,
  invChol_j,
  Vsub,
  nug_i,
  nug_j,
  df1,
  df2
)

crosspart_worker_R(xxi, xxj, xxi0, xxj0, invChol_i, invChol_j, Vsub, df1, df2)
}
\arguments{
\item{xxi}{numeric matrix xx from  partition i}

\item{xxj}{numeric matrix xx from  partition j}

\item{xxi0}{numeric matrix xx0 from  partition i}

\item{xxj0}{numeric matrix xx0 from  partition j}

\item{invChol_i}{numeric matrix invcholV from  partition i}

\item{invChol_j}{numeric matrix invcholV from  partition j}

\item{Vsub}{numeric variance matrix for Xij (upper block)}

\item{nug_i}{nugget from partition i}

\item{nug_j}{nugget from partition j}

\item{df1}{first degree of freedom}

\item{df2}{second degree of freedom}
}
\value{
a list of cross-partition statistics
}
\description{
Worker function 2 for partitioned GLS

R version of cross-part worker
}
\details{
Cross-partition statistics are calculated for a pair of partitions
i and j.

\code{crosspart_worker_R} is not exported by default and is just
a reference function. is a purely R version and will be removed in future
implementations.
}
\examples{
#TBA

}
\seealso{
Other remoteGLS: 
\code{\link{GLS_worker}()}
}
\concept{remoteGLS}
