% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{crosspart_worker_cpp}
\alias{crosspart_worker_cpp}
\title{worker function 2 for partitioned GLS}
\usage{
crosspart_worker_cpp(Li, Lj, Vij, df1, df2)
}
\arguments{
\item{Li}{list whose elements are those of a starmod.gls objects (i.e. created with
`GLS_worker_cpp()`). These are results from a single partition.}

\item{Lj}{list with same structure as Li that will be compared with Li}

\item{Vij}{numeric variance matrix for Xij}

\item{df1}{first degree of freedom}

\item{df2}{second degree of freedom}
}
\description{
worker function 2 for partitioned GLS
}
\details{
this function will also be used to perform the partitioned
version of the GLS in parallel and will compute cross-partition statistics
}
\examples{
#TBA
}
