% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_partition.R
\name{dist_km}
\alias{dist_km}
\title{Compute distance matrix in kilometers}
\usage{
dist_km(coords, coords2 = NULL)
}
\arguments{
\item{coords}{coordinate matrix}

\item{coords2}{optional coordinate matrix}
}
\value{
distance matrix
}
\description{
Compute distance matrix in kilometers
}
\details{
this function is simply a wrapper for \code{geosphere::distm()}
}
\seealso{
[geosphere::distm()]
}
