% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{distm_km}
\alias{distm_km}
\title{Calculate distance matrix in kilometers}
\usage{
distm_km(coords, coords2 = NULL)
}
\arguments{
\item{coords}{coordinate matrix}

\item{coords2}{optional coordinate matrix}
}
\value{
distance matrix
}
\description{
Calculate distance matrix in kilometers
}
\details{
this function is simply a wrapper for \code{geosphere::distm()}
}
\seealso{
[geosphere::distm()]

Other distm: 
\code{\link{distm_scaled}()}
}
\concept{distm}
