% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-functions.R
\name{distm_scaled}
\alias{distm_scaled}
\title{Calculate a unit distance matrix}
\usage{
distm_scaled(coords, coords2 = NULL, distm_FUN = "distm_km")
}
\arguments{
\item{coords}{a coordinate matrix with 2 columns and rows corresponding to
each location.}

\item{coords2}{an optional coordinate matrix}

\item{distm_FUN}{function used to calculate the distance matrix. This function
dictates the units of "max.dist"}
}
\value{
a distance matrix, scaled between 0 and 1, with a `max.dist`
attribute containing the maximum distance by which the initial values were
scaled.
}
\description{
Calculate a unit distance matrix
}
\details{
a distance matrix is fit and then divided by its maximum distance.
}
\examples{
map.width = 3 # square map width
coords = expand.grid(x = 1:map.width, y = 1:map.width) # coordinate matrix
distm_scaled(coords) # calculate distance matrix

}
\seealso{
Other distm: 
\code{\link{distm_km}()}
}
\concept{distm}
