% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.fitGLS2_cpp}
\alias{.fitGLS2_cpp}
\title{Fit GLS to remote sensing data}
\usage{
.fitGLS2_cpp(L, X, V, y, X0, nugget, save_xx, LL_only, no_F, threads)
}
\arguments{
\item{L}{empty remoteGLS list to be filled}

\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{y}{numeric vector}

\item{X0}{numeric matrix}

\item{nugget}{numeric nugget to add to V}

\item{save_xx}{logical: should xx, xx0, and invcholV be returned? This
functionality is meant for use with the partitioned GLS whereby these
values are used to calculate cross-partition statistics.}

\item{LL_only}{logical: should only the log-liklihood be computed?}

\item{no_F}{logical: should calculations needed for F tests be skipped?}

\item{threads}{number of threads used by Eigen for matrix algebra}
}
\description{
Fit GLS to remote sensing data
}
\details{
see `fitGLS()`
}
