% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.fitGLS_cpp}
\alias{.fitGLS_cpp}
\title{Fit GLS to remote sensing data}
\usage{
.fitGLS_cpp(X, V, y, X0, nugget, save_xx, threads)
}
\arguments{
\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{y}{numeric vector}

\item{X0}{numeric matrix}

\item{nugget}{numeric nugget to add to V}

\item{save_xx}{logical: should xx, xx0, and invcholV be returned? This
functionality is meant for use with the partitioned GLS whereby these
values are used to calculate cross-partition statistics.}

\item{threads}{number of threads used by Eigen for matrix algebra}
}
\description{
Fit GLS to remote sensing data
}
\details{
see `fitGLS()`
}
