% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.invchol_cpp}
\alias{.invchol_cpp}
\title{Invert the cholesky decomposition of V}
\usage{
.invchol_cpp(V, nugget = 0)
}
\arguments{
\item{V}{numeric matrix}

\item{nugget}{numeric nugget to add to variance matrix}
}
\description{
Invert the cholesky decomposition of V
}
\examples{
#TBA
}
