% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.optimize_nugget_cpp}
\alias{.optimize_nugget_cpp}
\title{Find the maximum likelihood estimate of the nugget}
\usage{
.optimize_nugget_cpp(X, V, y, lower, upper, tol, debug)
}
\arguments{
\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{y}{numeric vector}

\item{lower}{lower boundary for nugget search}

\item{upper}{upper boundary for nugget search}

\item{tol}{desired accuracy of nugget search}

\item{debug}{logical: debug mode?}
}
\description{
Find the maximum likelihood estimate of the nugget
}
\details{
this is the C++ version of `optimize()` which is specific to
finding the nugget value that maximizes the log-likelihood of `fitGLS_cpp()`

This function is a translation from the forchan algorithm fmin into C++:
http://www.netlib.org/fmm/fmin.f

Note: this function actually uses `LogLikGLS_cpp()` which should be swapped
for `fitGLS_cpp()` once the correct functionality is added to the latter.
}
\examples{
#TBA
}
