% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitSpatCor.R
\name{fit_spatialcor}
\alias{fit_spatialcor}
\title{Title}
\usage{
fit_spatialcor(
  X,
  t,
  r.start = 0.1,
  a.start = 1,
  fit.n = 100,
  fun = "exp",
  dist,
  location,
  scale.dist = TRUE,
  dist.scl = 1000,
  U = NULL,
  covars = NULL,
  plot.fig = FALSE,
  cols.plot = NULL,
  ...
)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Title
}
