% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_reml.R
\name{fitAR.map}
\alias{fitAR.map}
\title{Fit AR REML models to a time series matrix}
\usage{
fitAR.map(
  X,
  t,
  Z = NULL,
  ret_int.coef = FALSE,
  ret_AR.par = TRUE,
  ret_MSE = TRUE,
  ret_resid = TRUE,
  ret_logLik = TRUE
)
}
\arguments{
\item{X}{nxp time series response matrix with p columns corresponding to time
points and n columns corresponding to the number of pixels}

\item{t}{p length temporal response vector}

\item{ret_int.coef}{should the intercept coeffients be returned? logical}

\item{ret_AR.par}{should the AR parameter estimates be returned? logical}

\item{ret_MSE}{should the model MSEs be returned? logical}

\item{ret_resid}{should the model residuals be returned? logical}

\item{ret_logLik}{should the model log-likelihoods be returned? logical}
}
\value{
a list with the following elements: the initial function call
(\code{$call}), a coefficient matrix for the temporal variable
(\code{$time.coef}), an optional vector of the AR parameters (\code{$AR.par}),
an optional vector of MSEs (\code{$MSE}), an optional vector of
log-likelihoods (\code{$logLik}), and an optional nxp matrix of model
residuals (\code{$resids}).
}
\description{
Fit AR REML models to a time series matrix
}
\details{
by default the print.remoteAR() method does not show all output.
to access individual components, use \code{names()} to see element names
and the S3 \code{$} operator to access them.
}
\examples{
t = 1:30; n.pix = 10
X = matrix(rnorm(length(t)*n.pix), ncol = length(t))
fitAR.map(X, t) # only $call and $time.coef are printed by print.remoteAR()
summary(fitAR.map(X, t))

coef(fitAR.map(X, t)) # data frame of time coefficeints (alternatively fitAR.map(X, t)$time.coef)
fitAR.map(X, t)$AR.par # AR parameters
fitAR.map(X, t)$MSE # model MSEs
fitAR.map(X, t)$logLik # model log-likelihoods
resid(fitAR.map(X, t)) # matrix of model residuals (alternatively fitAR.map(X, t)$resids)
}
