% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLS_functions.R
\name{fitCLS.map}
\alias{fitCLS.map}
\title{Title}
\usage{
fitCLS.map(
  X,
  t,
  ret_xi.coef = FALSE,
  ret_int.coef = FALSE,
  ret_MSE = TRUE,
  ret_resid = TRUE
)
}
\arguments{
\item{ret_resid}{}
}
\value{

}
\description{
Title
}
