% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{fitGLS.partition_rcpp}
\alias{fitGLS.partition_rcpp}
\title{Title}
\usage{
fitGLS.partition_rcpp(
  X,
  y,
  X0,
  Dist,
  spatcor,
  Vfit.fun = "exponential-power",
  npart = 5,
  mincross = 4,
  nug.int = c(0, 1),
  nug.tol = 1e-05
)
}
\arguments{
\item{nug.tol}{}
}
\value{

}
\description{
Title
}
