% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS2.R
\name{fitGLS2}
\alias{fitGLS2}
\title{fit a remotePARTS GLS model}
\usage{
fitGLS2(
  formula,
  data,
  V,
  nugget = 0,
  form.0 = NULL,
  save_xx = FALSE,
  threads = 1,
  contrasts = NULL,
  ...
)
}
\arguments{
\item{formula}{formula to build the model with}

\item{data}{object containing the data}

\item{V}{Variance matrix}

\item{nugget}{nugget}

\item{form.0}{null model formula (default: "y ~ 1")}

\item{save_xx}{should xx be saved?}

\item{threads}{how many threads (functionality not yet available)}

\item{contrasts}{optional linear contrasts to use}

\item{...}{additional arguments passed to \code{optimize_nugget}}
}
\value{
a remoteGLS object
}
\description{
fit a remotePARTS GLS model
}
\examples{
TBA
}
