% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS_functions.R
\name{fitGLS_R}
\alias{fitGLS_R}
\title{fit GLS to remote sensing data}
\usage{
fitGLS_R(X, V, y, X0 = NULL, nugget = 0)
}
\arguments{
\item{X}{n x p numeric design matrix for predictor variables}

\item{V}{n x n numeric covariance matrix}

\item{y}{length n numeric resposne vector}

\item{X0}{n x p0 null numeric design matrix}

\item{nugget}{nugget to be added to variance matrix. see `?invert_cholR()`}
}
\value{
list containing relevant output paramters from model fit

Note: this output should be converted to an S3 class and values should be
extracted by defining `print.method()` and other methods.
}
\description{
fit GLS to remote sensing data
}
\examples{
#TBA
}
