% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{fitNugget_Rcpp}
\alias{fitNugget_Rcpp}
\title{Obtain the maximum likelihood nugget estimate from a `starmod.gls.cpp` model}
\usage{
fitNugget_Rcpp(X, V, y, int = c(0, 1), tol = 1e-05)
}
\arguments{
\item{X}{n x p numeric design matrix for predictor variables}

\item{V}{n x n numeric covariance matrix}

\item{y}{length n numeric resposne vector}

\item{int}{length 2 numeric vector specifying the inverval over which to
search of the MLE estimator. Default = `c(0, 1)`.}

\item{tol}{the desired accuracy of the mathematical optimization.
see `?optimize()`.}

\item{X0}{n x p0 null numeric design matrix}
}
\value{
maximum likelihood estimate of the nugget
}
\description{
Obtain the maximum likelihood nugget estimate from a `starmod.gls.cpp` model
}
\details{
this function is a wrapper that calls the C++ version of fitGLS()
and can be faster than fitNugget(). However, this functionallity is also
available in the C++ version of fitGLS()
}
