% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitV.R
\name{fitV}
\alias{fitV}
\alias{fitV.switch}
\title{Fit varcov matrix from distance matrix}
\usage{
fitV(Dist, spatialcor, method = "exponential")

fitV.switch(Dist, spatialcor, method = "exponential")
}
\arguments{
\item{Dist}{n x n numeric distance matrix}

\item{spatialcor}{spatial correlation parameter(s)}

\item{method}{function with which to transform the distance matrix into the
varcov matrix}
}
\value{
n x n variance-covariance matrix
}
\description{
Fit varcov matrix from distance matrix

Fit varcov matrix from distance matrix
}
\details{
\code{Dist} is transformed via \code{method}

\code{fitV.switch()} is not exported and is the same as
\code{fitV()} but uses \code{switch()} instead of \code{if else} statements.
\code{fitV.switch()} will likely be removed in future implementations.
}
\examples{

}
