% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{get_fm}
\alias{get_fm}
\title{generic method to extract fitted model from a complex object}
\usage{
get_fm(object, ...)
}
\arguments{
\item{object}{complex S3 object}

\item{...}{additional arguments passed to methods}
}
\description{
generic method to extract fitted model from a complex object
}
\seealso{
[fitCLS()] for example usage
}
