% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{get_fm.remoteCLS}
\alias{get_fm.remoteCLS}
\title{get linear model object from remote CLS}
\usage{
\method{get_fm}{remoteCLS}(obj)
}
\arguments{
\item{obj}{remoteCLS object}
}
\value{
lm class object from stats::lm
}
\description{
get linear model object from remote CLS
}
