% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{GLS.fit}
\alias{GLS.fit}
\title{Fit a GLS to remoteSTAR data}
\usage{
GLS.fit(formula, formula0 = NULL, data, V = NULL, invcholV = NULL)
}
\arguments{
\item{invcholV}{}
}
\value{

}
\description{
GLS.fit fits a GLS to the data given a specified V or invcholV
}
