% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{GLS.partition.data}
\alias{GLS.partition.data}
\title{Fit a GLS to specified partitions}
\usage{
GLS.partition.data(
  formula,
  formula0 = NULL,
  data,
  r,
  est.nugget = T,
  npart = 10,
  partition = NULL,
  fixed.nugget = NULL,
  nugget.tol = 1e-05,
  min.num.rSS = 12
)
}
\arguments{
\item{min.num.rSS}{}
}
\value{

}
\description{
GLS.partition.data fits a GLS to a specified number of random partitions of
the data after fitting the nugget. It calls GLS.fit().
}
