% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{GLS.partition.pvalue}
\alias{GLS.partition.pvalue}
\title{Wrapper for bootsrap test}
\usage{
GLS.partition.pvalue(part.out, nboot = 2000)
}
\arguments{
\item{part.out}{output of partitioned GLS model...}

\item{nboot}{number of bootstraps to run for F-test (NA skips bootstrap)}
}
\value{
list of p values calculated with different methods
}
\description{
Wrapper for bootsrap test
}
