% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteTS-mapTS_methods.R
\name{logLik.mapTS}
\alias{logLik.mapTS}
\title{Log-likelihood of mapTS object}
\usage{
\method{logLik}{mapTS}(object, ...)
}
\arguments{
\item{object}{mapTS object}

\item{...}{additional arguments to pass to the generic method}
}
\description{
Log-likelihood of mapTS object
}
