% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteTS-mapTS_methods.R
\name{logLik.remoteTS}
\alias{logLik.remoteTS}
\title{Log likelihood of remoteTS objects}
\usage{
\method{logLik}{remoteTS}(object, ...)
}
\arguments{
\item{object}{the remoteTS object}

\item{...}{additional arguments passed to the generic logLik function}
}
\description{
Log likelihood of remoteTS objects
}
