% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{LogLikGLS_cpp}
\alias{LogLikGLS_cpp}
\title{Caculate log-liklihood of GLS model}
\usage{
LogLikGLS_cpp(nugget, X, V, y)
}
\arguments{
\item{nugget}{the nugget to add to V}

\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{y}{numeric vector}
}
\description{
Caculate log-liklihood of GLS model
}
\details{
this function is mostly meant to optimize the nugget for a paritular
set of data.

Note: this function should be deprecated and simply added as functionality
to `fitGLS_cpp()`.
}
\examples{
#TBA
}
