% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{LS.fit}
\alias{LS.fit}
\title{Fit a LS model to a matrix of time series}
\usage{
LS.fit(X, t.scale)
}
\arguments{
\item{t.scale}{}
}
\value{

}
\description{
LS.fit fits a LS model using lm() to a matrix of time series located in
each row
}
