% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MatMult}
\alias{MatMult}
\title{Matrix multiplication, possibly with multiple cores}
\usage{
MatMult(A, B, cores)
}
\arguments{
\item{A}{numeric matrix}

\item{B}{numeric matrix}

\item{cores}{integer number of cores to use}
}
\description{
Matrix multiplication, possibly with multiple cores
}
