% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi}
\alias{ndvi}
\title{NDVI Remote sensing data for North America}
\format{
data frame with 348137 rows corresponding to sites and 37 columns:
\describe{
  \item{row}{grid row of the pixel}
  \item{col}{grid column of the pixel}
  \item{lng}{longitude of the pixel}
  \item{lat}{latitude of the pixel}
  \item{land}{dominant land class of the pixel}
  \item{ndvi<t>}{ndvi value of the pixel during the year <t>}
}
}
\usage{
ndvi
}
\description{
NDVI data for sites over time
}
\keyword{datasets}
