% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi_AK}
\alias{ndvi_AK}
\title{NDVI Remote sensing data for Alaska}
\format{
data frame with 31486 rows corresponding to sites and 38 columns:
\describe{
  \item{lng}{longitude of the pixel}
  \item{lat}{latitude of the pixel}
  \item{AR_coef}{pre-calculated AR REML coefficient}
  \item{CLS_coef}{pre-calculated CLS coefficient}
  \item{land}{dominant land class of the pixel}
  \item{land}{logical: is this land class rare?}
  \item{ndvi<t>}{ndvi value of the pixel during the year <t>}
}
}
\usage{
ndvi_AK
}
\description{
subset of the \code{ndvi} dataset containing only Alaska
}
\keyword{datasets}
