% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ndvi_AK10000}
\alias{ndvi_AK10000}
\title{NDVI remote sensing data for 10,000 random pixels from Alaska, with rare
land classes removed.}
\format{
data frame with 10,000 rows corresponding to sites and 37 columns:
\describe{
  \item{lng}{longitude of the pixel}
  \item{lat}{latitude of the pixel}
  \item{AR_coef}{pre-calculated AR REML coefficient standardized by mean ndvi
  values for each pixel}
  \item{CLS_coef}{pre-calculated CLS coefficient standardized by mean ndvi
  values for each pixel}
  \item{land}{dominant land class of the pixel}
  \item{land}{logical: is this land class rare?}
  \item{ndvi<t>}{ndvi value of the pixel during the year <t>}
}
}
\usage{
ndvi_AK10000
}
\description{
NDVI remote sensing data for 10,000 random pixels from Alaska, with rare
land classes removed.
}
\keyword{datasets}
