% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{nugget.fit}
\alias{nugget.fit}
\alias{nugget.fit.funct}
\title{fit a nugget to the covariance matrix V}
\usage{
nugget.fit.funct(nugget, formula, data, V, verbose = FALSE)

nugget.fit(
  formula,
  data,
  V,
  nugget.tol = 1e-05,
  interval = c(0, 1),
  verbose = FALSE
)
}
\arguments{
\item{verbose}{}

\item{interval}{}
}
\value{

}
\description{
nugget.fit fits a nugget to the covariance matrix V. Note that this involves
refitting the data using GLS.fit. nugget.fit uses the function nuget.fit.funct.
}
