% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_nugget.R
\name{optimize_nugget}
\alias{optimize_nugget}
\alias{fitNugget}
\alias{fitNugget_Rcpp}
\title{Find the maximum likelihood estimate of the nugget}
\usage{
optimize_nugget(X, V, y, lower = 0, upper = 1, tol = 1e-05, debug = FALSE)

fitNugget(X, V, y, lower = 0, upper = 1, tol = 1e-05)

fitNugget_Rcpp(X, V, y, lower = 0, upper = 1, tol = 1e-05)
}
\arguments{
\item{X}{numeric (double) nxp matrix}

\item{V}{numeric (double) nxn matrix}

\item{y}{numeric (double) nx1 column vector}

\item{lower}{lower boundary for nugget search}

\item{upper}{upper boundary for nugget search}

\item{tol}{desired accuracy of nugget search}

\item{debug}{logical: debug mode?}
}
\value{
maximum likelihood nugget estimate
}
\description{
Find the maximum likelihood estimate of the nugget

fitNugget - R version

fitNugget_Rcpp - Rcpp version
}
\details{
Finds the maximum likelihood nugget estimate via mathematical
optimization.

To maximize efficiency, \code{optimize_nugget()} is implemented entirely
in C++. Optimization takes place via a C++ version of the \code{fmin} routine
(Forsythe et al 1977). Translated from http://www.netlib.org/fmm/fmin.f

The function \code{LogLikGLS()} is optimized for \code{nugget}. Once the
\code{LogLikGLS()} functionality is absorbed by \code{fitGLS()}, it will
be used instead.

\code{fitNugget} is the R-only version of \code{optimize_nugget()}.
It uses \code{fitGLS_R()} and \code{stats::optimize()} to obtain the ML
nugget.

\code{fitNugget_Rcpp} is a hybrid between \code{optimize_nugget()}
and \code{fitNugget()}. It optimizes the C++ function \code{LogLikGLS()} with
the R function \code{stats::optimize()}.

After further testing, only the most efficeint of the 3 nugget optimizers will
remain in the package.
}
\examples{
#TBA

}
\seealso{
[stats::optimize()]
}
