% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{optimizeNugget_cpp}
\alias{optimizeNugget_cpp}
\title{Find the maximum likelihood estimate of the nugget}
\usage{
optimizeNugget_cpp(X, V, y, lower = 0, upper = 1, tol = 1e-05, debug = FALSE)
}
\arguments{
\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{y}{numeric vector}

\item{lower}{lower boundary for nugget search}

\item{upper}{upper boundary for nugget search}

\item{tol}{desired accuracy of nugget search}

\item{debug}{logical: debug mode?}
}
\description{
Find the maximum likelihood estimate of the nugget
}
\details{
this is the C++ version of `optimize()` which is specific to
finding the nugget value that maximizes the log-liklihood of `fitGLS_cpp()`

Note: this function actually uses `LogLikGLS_cpp()` which should be swapped
for `fitGLS_cpp()` once the correct funcionality is added.
}
\examples{
#TBA
}
