% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_partition.R
\name{part_csv}
\alias{part_csv}
\title{Extract a partition from a csv file}
\usage{
part_csv(part.i, csv.path, part.mat)
}
\arguments{
\item{part.i}{which partition to extract (integer)}

\item{csv.path}{path to csv file containing data}

\item{part.mat}{a matrix whose columns contain indices for a partition.}
}
\value{
a list with the following elements:

\describe{
    \item{\code{$y}}{the response vector of length \code{partsize}
    corresponding to parition i}
    \item{\code{$X}}{a model matrix for parition i}
    \item{\code{$coords}}{a matrix or data frame of spatial coordinates for
    partition i. The first column are x coordinates and the second are y
    coordinates}
}
}
\description{
Extract a partition from a csv file
}
\details{
\code{part_csv()} is meant to be used with \code{fitGLS.partition()}
}
\examples{
n.pix = 30865 # pixels in AK_ndvi_common-land.csv
parts = sample_partitions(npix = n.pix, npart = 4, partsize = 1000)

data.file = system.file("extdata", "AK_ndvi_common-land.csv",
                        package = "remotePARTS")

A.out = part_csv(part.i = 1, csv.path = data.file, part.mat = parts)

# dimensions of A.out elements:
dim(A.out$V)
dim(A.out$X)
dim(A.out$X0)
length(A.out$y)
}
\seealso{
[fitGLS.partition()]
}
