% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partGLS_ndviAK}
\alias{partGLS_ndviAK}
\title{partitioned GLS results}
\format{
an S3 class "partGLS" object. See ?fitGLS_partition() for further
details
}
\usage{
partGLS_ndviAK
}
\description{
Example output from fitGLS_partition() fit to the \code{ndvi_AK} data set
}
\keyword{datasets}
