% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLS_methods.R
\name{print.partGLS}
\alias{print.partGLS}
\title{S3 print method for "partGLS" objects}
\usage{
\method{print}{partGLS}(x, ...)
}
\arguments{
\item{x}{"partGLS" object}

\item{...}{additional arguments passed to print}
}
\description{
S3 print method for "partGLS" objects
}
