% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{print.remoteAR}
\alias{print.remoteAR}
\title{print method for remoteAR object}
\usage{
\method{print}{remoteAR}(x, ...)
}
\arguments{
\item{x}{remoteAR object}

\item{...}{additional arguments}
}
\value{
print formatted remoteAR
}
\description{
print method for remoteAR object
}
