% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{print.remoteCLS}
\alias{print.remoteCLS}
\title{print remoteCLS object}
\usage{
\method{print}{remoteCLS}(obj)
}
\arguments{
\item{obj}{remoteCLS object}
}
\description{
print remoteCLS object
}
