% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCor.R
\name{print.remoteCor}
\alias{print.remoteCor}
\title{S3 print method for "remoteCor" class}
\usage{
\method{print}{remoteCor}(x, ...)
}
\arguments{
\item{x}{remoteCor object to print}

\item{...}{additional arguments passed to print()}
}
\description{
S3 print method for "remoteCor" class
}
