% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{print.remoteGLS}
\alias{print.remoteGLS}
\title{print method for remoteGLS}
\usage{
\method{print}{remoteGLS}(obj, print.call = TRUE)
}
\arguments{
\item{obj}{remoteGLS object}
}
\value{
formatted output for remoteGLS object
}
\description{
print method for remoteGLS
}
