% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{remoteGLS}
\alias{remoteGLS}
\title{remoteGLS constructor (S3)}
\usage{
remoteGLS(form)
}
\value{
an empty S3 object of class "remoteGLS"
}
\description{
remoteGLS constructor (S3)
}
\examples{
tmp <- remoteGLS() #empty remoteGLS object
}
