% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotePARTS-package.R
\docType{package}
\name{remotePARTS-package}
\alias{remotePARTS}
\alias{remotePARTS-package}
\title{remotePARTS: Spatio-temporal Auto-regression for Remote Sensing Data}
\description{
"remoteSTAR is a set of tools for analyzing sets of spatially 
  correlated time series data, collected via remote sensing methods."
}
\section{remotePARTS functions}{

\code{\link{fitGLS}}
}

\author{
\strong{Maintainer}: Clay Morrow \email{morrow5@wisc.edu} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Anthony Ives \email{arives@wisc.edu} (\href{https://orcid.org/...}{ORCID})
}

}
\keyword{internal}
