% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteSTAR.R
\docType{package}
\name{remoteSTAR}
\alias{remoteSTAR}
\title{remoteSTAR}
\description{
Spatial-Temporal Auto-Regression of remote sensing data using spatial
spatial collapsing methods of Ives et al (unpublished).
}
\author{
Clay Morrow and Anthony Ives
}
