% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{residuals.remoteCLS}
\alias{residuals.remoteCLS}
\title{resisduals of remoteCLS object}
\usage{
\method{residuals}{remoteCLS}(obj)
}
\arguments{
\item{obj}{remoteCLS object}
}
\value{
model residuals
}
\description{
resisduals of remoteCLS object
}
