% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_partition.R
\name{sample_partitions}
\alias{sample_partitions}
\title{function to calculate partition size or number of partitions}
\usage{
sample_partitions(npix, pixels = NA, npart = 10, partsize = NA, verbose = TRUE)
}
\arguments{
\item{npix}{number of pixels in full dataset}

\item{pixels}{vector of pixel indexes to sample from}
}
\description{
function to calculate partition size or number of partitions
}
\examples{
dat.M <- matrix(rnorm(3000*20), ncol = 20)
dat.rast <- raster::raster(dat.M)

}
