% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{simX}
\alias{simX}
\title{Simulate a collection of correlated time series}
\usage{
simX(
  formula,
  data = data.frame(rep(1, n)),
  coef,
  b,
  s,
  Dr = NULL,
  t.scale,
  n,
  n.obs,
  n.burn,
  seed = 0
)
}
\arguments{
\item{seed}{}
}
\value{

}
\description{
simX will simulate ... [Description]
}
