% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{spatialcor.fit}
\alias{spatialcor.fit}
\alias{spatialcor.fit.data}
\title{Fit a distance matrix to a set of time series using CLS}
\usage{
spatialcor.fit(
  X,
  t.scale,
  Dist,
  r.start = 0.1,
  fit.n.sample,
  FUN = "exponential",
  plot.fig = F,
  col.plot = NULL
)

spatialcor.fit.data(
  X,
  t.scale,
  data,
  r.start = 0.1,
  fit.n.sample,
  FUN = "exponential",
  plot.fig = F,
  col.plot = NULL
)
}
\arguments{
\item{col.plot}{}

\item{data}{}
}
\value{

}
\description{
spatialcor.fit fits an exponential or taper-spherical distance matrix to a
spatial set of time series by performing CLS, calculating the correlation
matrix of residuals, and then fitting the correlations using nls.

spatialcor.fit.data fits an exponential or taper-spherical distance matrix
to a spatial set of time series by performing CLS, calculating the
correlation matrix of residuals, and then fitting the correlations using nls.

NOTE: In contrast to spatialcor.fit, spatialcor.fit.data returns the spatial
correlation scaled to the input distance matrix in km (from distm)
}
