% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLS_methods.R
\name{t.test.partGLS}
\alias{t.test.partGLS}
\title{Conduct a t-test of "partGLS" object}
\usage{
\method{t.test}{partGLS}(x, ...)
}
\arguments{
\item{x}{"partGLS" object}

\item{...}{additional arguments passed to print}
}
\value{
a coefficient table with estimates, standard errors, t-statistics, and p-values
}
\description{
Conduct a correlated t-test of a partitioned GLS
}
