% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{taper.spherical}
\alias{taper.spherical}
\alias{taper.spherical.dif}
\title{Create tapered distance matrices}
\usage{
taper.spherical(d, beta)

taper.spherical.dif(d, cor, b)
}
\value{

}
\description{
taper.spherical and taper.spherical.dif are different formulations for
tapered distance matrices in which cells more distant than beta are set to
zero.
}
\examples{
taper.spherical(d, beta)
taper.spherical.dif(d, cor, b)
}
