% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitCor.R
\name{test_covar_fun}
\alias{test_covar_fun}
\title{Test passing a covariance function and arguments}
\usage{
test_covar_fun(d, covar_FUN = "covar_exppow", covar.pars = list(range = 0.5))
}
\arguments{
\item{d}{numeric vector or matrix of distances}

\item{covar_FUN}{distance-based covariance function to use,
which must take \code{d} as its first argument}

\item{covar.pars}{vector or list of parameters (other than d) passed to the
covar function}
}
\description{
Test passing a covariance function and arguments
}
