% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tinvchol_cpp}
\alias{tinvchol_cpp}
\title{Find the transposed inverse cholesky decomposition of V}
\usage{
tinvchol_cpp(V, nugget = 0)
}
\arguments{
\item{V}{numeric matrix}

\item{nugget}{numeric nugget to add to variance matrix}
}
\description{
Find the transposed inverse cholesky decomposition of V
}
\examples{
#TBA
}
