% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{V.fit}
\alias{V.fit}
\title{Construct covariance matrix for remoteSTAR structures}
\usage{
V.fit(Dist, spatialcor, FUN = "exponential")
}
\arguments{
\item{FUN}{}
}
\value{

}
\description{
V.fit constructs the covariance matrix from a distance matrix for exponential
and taper-spherical structures
}
