% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioned-GLS.R
\name{calc_dfpart}
\alias{calc_dfpart}
\title{calculate degrees of freedom for partitioned GLS}
\usage{
calc_dfpart(partsize, p, p0)
}
\arguments{
\item{partsize}{number of pixels in each partition}

\item{p}{number of predictors in alternate model}

\item{p0}{number of parameters in null model}
}
\value{
a named vector containing the first and second degrees of freedom ("df1" and "df2", respectively)
}
\description{
calculate degrees of freedom for partitioned GLS
}
