% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioned-GLS.R
\name{calc_dfpart}
\alias{calc_dfpart}
\title{calculate degrees of freedom for partitioned GLS}
\usage{
calc_dfpart(partsize, p, p0)
}
\arguments{
\item{partsize}{number of pixels in each partition}

\item{p}{number of predictors in alternate model}

\item{p0}{number of parameters in null model}
}
\description{
calculate degrees of freedom for partitioned GLS
}
\examples{
# calc_dfpart(partsize = 2000, p = 4, p0 = 1)

}
