% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{check_posdef}
\alias{check_posdef}
\title{check if matrix is positive definitive}
\usage{
check_posdef(M)
}
\arguments{
\item{M}{numeric matrix}
}
\description{
check if matrix is positive definitive
}
