% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{coef.remoteAR}
\alias{coef.remoteAR}
\title{Extract coefficients from remoteAR object}
\usage{
\method{coef}{remoteAR}(object, ...)
}
\arguments{
\item{object}{remoteAR object}

\item{...}{additional arguments}
}
\value{
coefficeint data frame
}
\description{
Extract coefficients from remoteAR object
}
