% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{cor_t.test}
\alias{cor_t.test}
\title{T test of partitioned GLS}
\usage{
cor_t.test(object)
}
\arguments{
\item{object}{remoteGLS.parts object}
}
\value{
t-table
}
\description{
T test of partitioned GLS
}
