% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS_functions.R
\name{correlated.F.bootstrap}
\alias{correlated.F.bootstrap}
\title{bootsrap test}
\usage{
correlated.F.bootstrap(Fmean.obs, rSSR, rSSE, df1, df2, npart, nboot = 2000)
}
\arguments{
\item{Fmean.obs}{Fmean}

\item{rSSR}{rSSR}

\item{rSSE}{rSSE}

\item{df1}{df1}

\item{df2}{df2}

\item{npart}{partitions}

\item{nboot}{bootstraps}
}
\value{

}
\description{
bootsrap test
}
\examples{
#TBA
}
