% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_reml.R
\name{fitAR}
\alias{fitAR}
\alias{AR_funct}
\title{Pixel AR REML}
\usage{
fitAR(formula, data)

AR_funct(par, x, U, LL.only = TRUE)
}
\arguments{
\item{formula}{model formula}

\item{data}{object in which the model will first look for data}

\item{par}{AR parameter value}

\item{x}{vector of time series (response)}

\item{U}{model matrix (predictors)}

\item{LL.only}{logical: should only the log-liklihood be computed}
}
\value{
\code{AR_funct()} will return only the log-likelihood
if \code{LL.only = TRUE}.

Otherwise, a \code{remoteAR.pixel} object is returned,
which is a list with the following elements:

\describe{
    \item{\code{$call}}{matched function call}
    \item{\code{$coef}}{regression coefficeint table}
    \item{\code{$b}}{REML AR parameter}
    \item{\code{$MSE}}{model MSE}
    \item{\code{$resids}}{model residuals}
    \item{\code{$s2beta}}{estimated coefficient covariance matrix}
    \item{\code{$logLik}}{model log-likelihood}
}
}
\description{
Fit an Auto-regressive time series analysis using restricted
maximum likelihood
}
\details{
\code{fitAR()} obtains the REML AR parameter by performing
mathematical optimization for \code{par} in \code{AR_funct(LL.only = TRUE)}.
Once the REML parameter is obtained, \code{AR_funct(LL.only = FALSE)} is
called.

By default, the print.remoteAR() method does not show all output.
to access individual components, use \code{names()} to see element names
and the S3 \code{$} operator to access them.
}
\examples{
time = 1:30
x = rnorm(31)
x = x[2:31] + x[1:30] + 0.3*time #AR(1) process + time trend
U = stats::model.matrix(formula(x ~ time))

# using the AR function with a given paramter
AR_funct(par = .2, x, U, LL.only = TRUE)
AR_funct(par = .2, x, U, LL.only = FALSE)

# find the REML solution for par and save results
AR <- fitAR(x ~ time)

# extract info
coef(AR) # time coefficient table; also AR$coef
AR$MSE # model MSE
AR$resids # model residuals
AR$b # AR parameter estimate
AR$logLik  # log-likelihood of the model
}
\seealso{
[fitAR.map()] for fitting to a full map and [fitCLS()] &
[fitCLS.map()] for CLS versions.
}
\concept{remoteAR}
