% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_reml.R
\name{fitAR}
\alias{fitAR}
\title{Fit an Auto-regressive time series analysis using restricted maximum
likelihood}
\usage{
fitAR(formula, data)
}
\arguments{
\item{formula}{model formula}

\item{data}{object in which the model will first look for data}
}
\value{
a "remoteAR.pixel" object. See [AR_funct()] for more details.
}
\description{
Fit an Auto-regressive time series analysis using restricted maximum
likelihood
}
\details{
[fitAR()] is a wrapper function for [AR_funct()].

By default, the print.remoteAR() method does not show all output.
to access individual components, use \code{names()} to see element names
and the S3 \code{$} operator to access them.
}
\examples{
time = 1:30
x = rnorm(31)
x = x[2:31] + x[1:30] + 0.3*time #AR(1) process + time trend
fitAR(x ~ time)
}
