% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLS_functions.R
\name{fitCLS}
\alias{fitCLS}
\title{Fit a CLS regression to a time series}
\usage{
fitCLS(x, t, Z = NULL, save_AR.df = FALSE)
}
\arguments{
\item{x}{numeric vector of length p containing time series data for one
location (i.e. a pixel)}

\item{t}{numeric vector of length p containing the values for time.}

\item{Z}{(optional) a numeric vector/matrix with p rows and q columns. Each
column contains a different covariate measured at the same site as x at the
same p time points.}

\item{save_AR.df}{should the auto-regression data frame be returned?
default: FALSE}
}
\value{
\code{remoteCLS} object. A list with the following elements:
\describe{
    \item{\code{$call}}{the matched call to this function}
    \item{\code{$fm}}{the model object fit using \code{stats::lm()}}
    \item{\code{$AR.df}}{an optionally returned AR data frame built with
    \code{AR_df()}}
}
}
\description{
Fit a CLS regression to a time series
}
\details{
CLS is fit to a time series \code{x} by first building an AR
data frame with \code{AR_df()}. Then, \code{x} at time j (\code{x.tj})
is regressed on \code{x} at time i (\code{x.ti}) and \code{t} at time j
(\code{t.j}).

By default the print.remoteCLS() method does not *show* all output,
even if the ret_* flags are set to TRUE. Use the S3 \code{$} operator to
access elements by name or use one of the extractor functions
(shown in examples below).
}
\examples{
data(ndvi_AK3000)
x = unlist(ndvi_AK3000[1, -c(1:6)]) # time series for first pixel only
t = 1:length(x) # time points

## CLS without covariates
fitCLS(x = x, t = t)

## now with 2 covariates
Z = cbind(rnorm(length(x)), rnorm(length(x)))
fitCLS(x = x, t = t, Z = Z)

## get the AR data frame with save_AR.df = TRUE
fitCLS(x = x, t = t, save_AR.df = TRUE)$AR.df

## print a summary
summary(fitCLS(x = x, t = t))

## use get_fm() to extract $fm, which is a stats::lm() object
fm = get_fm(fitCLS(x = x, t = t))
## then treat it just like lm()
summary(fm)
resid(fm)
}
\seealso{
[AR_df()] for AR data frames, [fitCLS.map()] for fitting full-map
time series CLS, [fitAR()] & [fitAR.map()] for using AR REML instead of CLS.
}
\concept{remoteCLS}
