% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CLS_functions.R
\name{fitCLS}
\alias{fitCLS}
\title{Fit a CLS regression to a time series}
\usage{
fitCLS(x, t, Z = NULL)
}
\arguments{
\item{x}{numeric vector of length p containing remote sensing data}

\item{t}{numeric vector of length p containing the values for time. Recommended:
`scale(1:ncol(X))`}

\item{Z}{(optional) a numeric vector/matrix with p columns and q rows. Each
row contains a different covariate measured at the same site as x at the same
p time points.}
}
\value{
a list of 3 or 4 elements:
`coef` the model coefficients from `summary(lm(...))$coef`
`MSE` means squared error of the regression model
`resids` the residual errors of the model
`Z` the coefficients for optional covariates
}
\description{
see `?cls_star()`
}
\examples{
#TBA
}
