% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{fitGLS.partition}
\alias{fitGLS.partition}
\title{Title}
\usage{
fitGLS.partition(
  X,
  V,
  y,
  X0,
  nugget = 0,
  npart = 10,
  mincross = 5,
  nug.int = c(0, 1),
  nug.tol = 1e-05
)
}
\arguments{
\item{nug.tol}{}
}
\value{

}
\description{
Title
}
