% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{fitGLS.partition}
\alias{fitGLS.partition}
\title{fit GLS model by partitioning remote sensing data}
\usage{
fitGLS.partition(
  X,
  V,
  y,
  X0,
  nugget = 0,
  npart = 10,
  mincross = 5,
  nug.int = c(0, 1),
  nug.tol = 1e-05
)
}
\arguments{
\item{X}{n x p numeric design matrix for predictor variables}

\item{V}{n x n numeric covariance matrix}

\item{y}{length n numeric resposne vector}

\item{X0}{n x p0 null numeric design matrix}

\item{nugget}{nugget to be added to variance matrix. see `?invert_cholR()`}

\item{npart}{integer: number of of partitions to divide the data into}

\item{mincross}{intiger: minimum number of partition pairs from which to
calculate statistics (i.e. )}

\item{nug.int}{interval of nugget passed to fitGLS()}

\item{nug.tol}{accuracy of nugget calculation passed to fitGLS()}
}
\value{
list of GLS statistics
}
\description{
fit GLS model by partitioning remote sensing data
}
\details{
Note: This function is not complete yet. Use the C++ version instead
}
\examples{
#TBA
}
