% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_partition.R
\name{fitGLS.partition_rcpp}
\alias{fitGLS.partition_rcpp}
\title{fit GLS model by partitioning remote sensing data via Rcpp}
\usage{
fitGLS.partition_rcpp(
  X,
  y,
  X0,
  Dist,
  spatcor,
  Vfit.fun = "exponential-power",
  npart = 5,
  mincross = 4,
  nug.int = c(0, 1),
  nug.tol = 1e-05,
  workerB_cpp = TRUE
)
}
\arguments{
\item{X}{n x p numeric design matrix of predictor variables}

\item{y}{length n numeric response vector}

\item{X0}{n x p0 null numeric design matrix}

\item{Dist}{distance matrix}

\item{spatcor}{spatial correlation parameter(s)}

\item{Vfit.fun}{function to use for Vfit() calculation}

\item{npart}{number of partitions}

\item{mincross}{minimum number of partition pairs from which to calculate
statistics}

\item{nug.int}{interval of nugget search}

\item{nug.tol}{accuracy of nugget estimate}

\item{workerB_cpp}{logical: should the cpp version of worker function be
used? this argument is deprecated and was just used to test}
}
\value{
list of GLS statistics
}
\description{
fit GLS model by partitioning remote sensing data via Rcpp
}
\details{
\code{fitGLS.partition_rcpp()} uses \code{GLS_worker()} to
obtain partition-specific statistics and then \code{crosspart_worker()}
to obtain cross-partition statistics.

Currently, \code{fitGLS.parition_rcpp()} needs the full distance
matrix \code{Dist}. \code{fitGLS.partition()} is in development and will
not require the full matrix be loaded into memory.
}
\examples{

}
