% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS_functions.R
\name{fitGLS.partition_rcpp}
\alias{fitGLS.partition_rcpp}
\title{fit GLS model by partitioning remote sensing data via Rcpp}
\usage{
fitGLS.partition_rcpp(
  X,
  y,
  X0,
  Dist,
  spatcor,
  Vfit.fun = "exponential-power",
  npart = 5,
  mincross = 4,
  nug.int = c(0, 1),
  nug.tol = 1e-05,
  workerB_cpp = TRUE
)
}
\arguments{
\item{X}{n x p numeric design matrix for predictor variables}

\item{y}{length n numeric resposne vector}

\item{X0}{n x p0 null numeric design matrix}

\item{Dist}{distance matrix}

\item{spatcor}{spatial correlation parameter(s)}

\item{Vfit.fun}{function to use for Vfit() calculation}

\item{npart}{number of partitions}

\item{mincross}{minimum number of parition pairs from which to calculate
statistics}

\item{nug.int}{interval of nugget search}

\item{nug.tol}{accuracy of nugget estimate}

\item{workerB_cpp}{logical: should the cpp version of worker function be
used? this argument is deprecated and was just used to test}
}
\value{
list of GLS statistics
}
\description{
fit GLS model by partitioning remote sensing data via Rcpp
}
\examples{
#TBA
}
