% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{fitGLS}
\alias{fitGLS}
\alias{LogLikGLS}
\title{Fit GLS to remote sensing data}
\usage{
fitGLS(X, V, y, X0, nugget = 0, save_xx = FALSE, threads = 1)

LogLikGLS(nugget, X, V, y)
}
\arguments{
\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{y}{numeric vector}

\item{X0}{numeric (double) matrix}

\item{nugget}{the nugget to add to V}

\item{save_xx}{logical: should xx, xx0, and invcholV be returned? This
functionality is meant for use with the partitioned GLS whereby these
values are used to calculate cross-partition statistics.}

\item{threads}{integer indicating the number of threads to use. This current
version does not have multi-thread functionality so this argument does
nothing yet.}
}
\description{
Fit GLS to remote sensing data

Caculate log-liklihood of GLS model
}
\details{
see `fitGLS()`

this function is mostly meant to optimize the nugget for a paritular
set of data.

Note: this function should be deprecated and simply added as functionality
to `.fitGLS_cpp()`.
}
\examples{
#TBA

#TBA
}
