% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS_opt.R
\name{fitGLS_opt_FUN}
\alias{fitGLS_opt_FUN}
\title{Function that fitGLS_opt optimizes over}
\usage{
fitGLS_opt_FUN(
  op,
  fp,
  formula,
  data = NULL,
  coords,
  covar_FUN = "covar_exp",
  distm_FUN = "distm_scaled"
)
}
\arguments{
\item{op}{a named vector of parameters to be optimized}

\item{fp}{a named vector of fixed parameters}

\item{formula}{GLS model formula}

\item{data}{data source}

\item{coords}{a coordinate matrix}

\item{covar_FUN}{a covariance function}

\item{distm_FUN}{a distm function}
}
\value{
\code{fitGLS_opt_FUN} returns the negative log likelihood of a GLS,
given the parameters in \code{op} and \code{fp}
}
\description{
Function that fitGLS_opt optimizes over
}
\examples{
\dontrun{
data(ndvi_AK10000)
df = ndvi_AK10000[seq_len(200), ] # first 500 rows
coords = df[, c("lng", "lat")]
remotePARTS:::fitGLS_opt_FUN(op = c(range = .1, nugget = .2),
                             formula = CLS_coef ~ 0 + land, data = df,
                             coords = coords)
remotePARTS:::fitGLS_opt_FUN(op = c(range = .1), fp = c(nugget = 0),
                             formula = CLS_coef ~ 0 + land, data = df,
                             coords = coords)
}
}
