% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{fitV}
\alias{fitV}
\title{Fit varcov matrix from distance matrix}
\usage{
fitV(Dist, spatialcor, fun = "exponential")
}
\arguments{
\item{Dist}{n x n numeric distance matrix}

\item{spatialcor}{spatial correlation parameter(s)}

\item{fun}{function with which to transform the distance matrix into the
varcov matrix}
}
\value{
n x n varcov matrix
}
\description{
Fit varcov matrix from distance matrix
}
\details{
Note: currently the "exponential" version yields a non-positive definitive
(and therefore invalid) V matrix.
}
\examples{
#TBA
}
