% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GLS_functions.R
\name{GLS.partition.pvalue}
\alias{GLS.partition.pvalue}
\title{Wrapper for bootsrap test}
\usage{
GLS.partition.pvalue(starpart, nboot = 2000)
}
\arguments{
\item{starpart}{output of partitioned GLS model...}

\item{nboot}{bootsraps}
}
\value{

}
\description{
Wrapper for bootsrap test
}
\examples{
#TBA
}
