% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remote-sensing-functions.R
\name{GLS.partition.pvalue}
\alias{GLS.partition.pvalue}
\title{Wrapper for correlated.F.bootstrap()}
\usage{
GLS.partition.pvalue(z, nboot = 2000)
}
\arguments{
\item{nboot}{}
}
\value{

}
\description{
GLS.partition.pvalue is a wrapper for correlated.F.bootstrap()
}
