% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GLS_worker_cpp}
\alias{GLS_worker_cpp}
\title{worker function 1 for paritioned GLS}
\usage{
GLS_worker_cpp(y, X, V, X0, save_xx = FALSE)
}
\arguments{
\item{y}{numeric vector}

\item{X}{numeric matrix}

\item{V}{numeric matrix}

\item{X0}{numeric matrix}

\item{save_xx}{logical: should xx, xx0, and tInvCholV be returned?}
}
\description{
worker function 1 for paritioned GLS
}
\details{
this function will eventually be used to perform the partitioned
version of the GLS in parallel
}
\examples{
#TBA
}
