% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_chol.R
\name{invert_chol}
\alias{invert_chol}
\title{Invert the cholesky decomposition of V}
\usage{
invert_chol(M, nugget = 0, debug = FALSE)
}
\arguments{
\item{M}{numeric matrix}

\item{nugget}{numeric nugget to add to V}

\item{debug}{logical: enter debug mode?}
}
\value{
numeric matrix: inverse of the cholesky decomposition
}
\description{
Invert the cholesky decomposition of V
}
\examples{
M <- crossprod(matrix(1:6, 3))
# without a nugget:
invert_chol(M)
# with a nugget:
invert_chol(M, nugget = 0.2)

}
