% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_GLS.R
\name{optim_GLS_func}
\alias{optim_GLS_func}
\alias{optimize_GLS}
\title{function to optimize GLS parameters r, a, and nugget}
\usage{
optim_GLS_func(
  in.pars = c(r = 0.01, a = 1, nug = 0.1),
  y,
  modmat,
  D,
  verbose = FALSE,
  V.meth = "exponential-power",
  constrain = FALSE
)

optimize_GLS(
  formula,
  coords,
  dist_f = "dist_km",
  V.meth = "exponential",
  nugget = NA,
  spcor = NA,
  verbose = FALSE,
  data,
  pars.start = c(r = 0.1, a = 1, nug = 0.2),
  save_xx = FALSE,
  ret.GLS = FALSE,
  algorithm = "Nelder-Mead",
  debug = FALSE,
  ...
)
}
\arguments{
\item{in.pars}{parameters to optimize}

\item{y}{response vector}

\item{modmat}{predictor model matrix}

\item{D}{distance matrix}

\item{verbose}{should additional info be printed?}

\item{V.meth}{covariance method for fitting V}

\item{constrain}{do parameters need to be constrained internally?}

\item{formula}{model formula}

\item{coords}{2-column matrix of x and y coordinates}

\item{dist_f}{distance function to use}

\item{nugget}{spatial nugget. If NA, a nugget will be estimated}

\item{spcor}{spatial correlation paramters either NA, a single value (r) or
a named vector: c(r = ..., a = ...). if spcor = NA, these parameters will be
estimated}

\item{data}{data object for which formula tries to match}

\item{pars.start}{default starting values for the spcor}

\item{save_xx}{save cross correlation stats?}

\item{ret.GLS}{return GLS?}

\item{algorithm}{method passed to optimizer (default = 'Nelder-Mead')
see \code{?optim()} for available methods}

\item{debug}{boolean: debug mode?}

\item{...}{additional arguments passed to \code{fitGLS2()}}
}
\value{
a remoteGLS object
}
\description{
function to optimize GLS parameters r, a, and nugget

Fit a GLS by estimating r, a, and nugget
}
\details{
When \code{constrain = FALSE}, it is assumed that parameter values
 given by \code{in.pars} are already constrained to their desired range. When
 When \code{constrain = TRUE}, parameter values are internally transformed.

 r and nugget are transformed with an inverse logit
 (i.e., \code{function(x){1 / (1 + exp(-x))}}) and \code{a} is transformed
 with an exponential function (i.e., exp(a)).
}
\seealso{
[fitGLS2()]
}
