% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_GLS.R
\name{optimize_GLS}
\alias{optimize_GLS}
\title{Fit a PARTS GLS model, with maximum likelihood spatial parameters}
\usage{
optimize_GLS(
  formula,
  data = NULL,
  coords,
  distm_FUN = "distm_scaled",
  covar_FUN = "covar_exp",
  start = c(range = 0.01, nugget = 0),
  fixed = c(),
  opt.only = FALSE,
  formula0 = NULL,
  save.xx = FALSE,
  save.invchol = FALSE,
  no.F = TRUE,
  ...
)
}
\arguments{
\item{formula}{a model formula, passed to \code{fitGLS}}

\item{data}{an optional data frame environment in which to search for
variables given by \code{formula}; passed to \code{fitGLS}}

\item{coords}{a numeric coordinate matrix or data frame, with two columns and
rows corresponding to each pixel}

\item{distm_FUN}{a function to calculate a distance matrix from \code{coords}}

\item{covar_FUN}{a function to estimate distance-based covariances}

\item{start}{a named vector of starting values for each parameter to be estimated;
names must match the names of arguments in \code{covar_FUN} or "nugget"}

\item{fixed}{an optional named vector of fixed parameter values; names
must match the names of arguments in \code{covar_FUN} or "nugget"}

\item{opt.only}{logical: if TRUE, execution will halt after estimating the parameters;
a final GLS will not be fit with the estimated parameters}

\item{formula0, save.xx, save.invchol, no.F}{arguments passed to \code{fitGLS}
for final GLS output}

\item{...}{additional arguments passed to \code{stats::optim()}}
}
\value{
If \code{opt.only = TRUE}, \code{optimize_GLS} returns the
output from \code{stats::optim()}: see it's documentation for more details.

Otherwise, a list with two elements is returned:

\describe{
    \item{opt}{output from \code{optim}, as above}
    \item{GLS}{a "remoteGLS" object. See \code{\link{fitGLS}} for more details.}
}
}
\description{
Fit a PARTS GLS model, with maximum likelihood spatial parameters
}
\details{
Estimate spatial parameters, via maximum likelihood, from data
rather than from time series residuals; Fit a GLS with these specifications.

\code{optimize_GLS} fits a GLS by estimating spatial parameters from
data. \code{\link{fitCor}}, combined with \code{\link{fitGLS}(nugget = NA)},
gives better estimates of spatial parameters, but time-series residuals may
not be available in all cases. In these cases, spatial parameters can be
estimated from distances among points and a response vector. Mathematical
optimization of the log likelihood of different GLS models are computed by
calling \code{optim()} on \code{fitGLS}.

Distances are calculated with \code{distm_FUN} and a covariance matrix is
calculated from these distances with \code{covar_FUN}. Arguments to to
\code{covar_FUN}, except distances, are given by \code{start} and \code{fixed}.
Parameters specified in \code{start} will be be estimated while those given
by \code{fixed} will remain constant throughout fitting. Parameter names in
\code{start} and \code{fixed} should exactly match the names of arguments in
\code{covar_FUN} and should not overlap (though, \code{fixed} takes precedence).

In addition to arguments of \code{covar_FUN} a "nugget" component can
also be occur in \code{start} or \code{fixed}. If "nugget" does not occur
in either vector, the GLS are fit with \code{nugget = 0}. A zero nugget also
allows much faster computation, through recycling the common
inverse cholesky matrix in each GLS computation. A non-zero nugget requires
inversion of a different matrix at each iteration, which can be
substantially slower.

If \code{opt.only = FALSE}, the estimated parameters are used to fit the final
maximum likelihood GLS solution with \code{fitGLS()} and arguments
\code{formula0}, \code{save.xx}, \code{save.invchol}, and \code{no.F}.

Some parameter combinations may not produce valid covariance matrices. During
the optimization step messages about non-positive definitive V may result on
some iterations. These warnings are produced by \code{fitGLS} and NA
log-likelihoods are returned in those cases.

Note that \code{optimize_GLS} fits multiple GLS models, which requires
inverting a large matrix for each one (unless a fixed 0 nugget is used).
This process is very computationally intensive and may take a long time to
finish depending upon your machine and the size of the data.
}
\examples{
## read data
data(ndvi_AK3000)
df = ndvi_AK3000[seq_len(1000), ] # first 1000 rows

## estimate nugget and range
optimize_GLS(formula = CLS_coef ~ 0 + land, data = df,
            coords = df[, c("lng", "lat")], start = c(range = .1, nugget = 0),
            opt.only = TRUE)

## estimate range only, fixed nugget at 0, and fit full GLS
optimize_GLS(formula = CLS_coef ~ 0 + land, data = df,
             coords = df[, c("lng", "lat")],
             start = c(range = .1), fixed = c("nugget" = 0),
             method = "Brent", lower = 0, upper = 1)

}
\seealso{
\code{\link{fitCor}} for estimating spatial parameters from time
series residuals; \code{\link{fitGLS}} for fitting GLS and with the option
of estimating the maximum-likelihood nugget component only.
}
