% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_nugget.R
\name{optimize_nugget}
\alias{optimize_nugget}
\title{Find the maximum likelihood estimate of the nugget}
\usage{
optimize_nugget(X, V, y, lower = 0, upper = 1, tol = 1e-05, debug = FALSE)
}
\arguments{
\item{X}{numeric (double) nxp matrix}

\item{V}{numeric (double) nxn matrix}

\item{y}{numeric (double) nx1 column vector}

\item{lower}{lower boundary for nugget search}

\item{upper}{upper boundary for nugget search}

\item{tol}{desired accuracy of nugget search}

\item{debug}{logical: debug mode?}
}
\description{
Find the maximum likelihood estimate of the nugget
}
\details{
this is the C++ version of `optimize()` which is specific to
finding the nugget value that maximizes the log-likelihood of `fitGLS_cpp()`

This function is a translation from the forchan algorithm fmin into C++:
http://www.netlib.org/fmm/fmin.f

Note: this function actually uses `LogLikGLS_cpp()` which should be swapped
for `fitGLS_cpp()` once the correct functionality is added to the latter.
}
\examples{
#TBA

}
