% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimize_nugget.R
\name{optimize_nugget}
\alias{optimize_nugget}
\title{Find the maximum likelihood estimate of the nugget}
\usage{
optimize_nugget(
  X,
  y,
  V,
  lower = 0.001,
  upper = 0.999,
  tol = .Machine$double.eps^0.25,
  debug = FALSE,
  ncores = NA
)
}
\arguments{
\item{X}{numeric (double) nxp matrix}

\item{y}{numeric (double) nx1 column vector}

\item{V}{numeric (double) nxn matrix}

\item{lower}{lower boundary for nugget search}

\item{upper}{upper boundary for nugget search}

\item{tol}{desired accuracy of nugget search}

\item{debug}{logical: debug mode?}

\item{ncores}{an optional integer indicating how many CPU threads to use for
matrix calculations.}
}
\value{
maximum likelihood nugget estimate
}
\description{
Find the maximum likelihood estimate of the nugget
}
\details{
Finds the maximum likelihood nugget estimate via mathematical
optimization.

To maximize efficiency, \code{optimize_nugget()} is implemented entirely
in C++. Optimization takes place via a C++ version of the \code{fmin} routine
(Forsythe et al 1977). Translated from http://www.netlib.org/fmm/fmin.f

The function \code{LogLikGLS()} is optimized for \code{nugget}. Once the
\code{LogLikGLS()} functionality is absorbed by \code{fitGLS()}, it will
be used instead.
}
\examples{
\dontrun{
## read data
data(ndvi_AK10000)
df = ndvi_AK10000[seq_len(200), ] # first 200 rows

## format data
X = stats::model.matrix(CLS_coef ~ 0 + land, data = df)

## fit covariance matrix
V = covar_exp(distm_scaled(cbind(df$lng, df$lat)), range = .01)

## find the ML nugget
remotePARTS:::optimize_nugget(X = X, V = V, y = df$CLS_coef, debug = TRUE)
}
}
\seealso{
\code{?stats::optimize()}
}
