% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLS_methods.R
\name{part_chisqr}
\alias{part_chisqr}
\title{Chisqr test for partitioned GLS}
\usage{
part_chisqr(Fmean, rSSR, df1, npart)
}
\arguments{
\item{Fmean}{mean value of F-statistic from correlated F-tests}

\item{df1}{first degree of freedom for F-tests}

\item{npart}{number of partitions}
}
\value{
a p-value for the correlated chisqr test
}
\description{
Chisqr test for partitioned GLS
}
\examples{
remotePARTS:::part_chisqr(Fmean = 3.6, rSSR = .021, df1 = 2, npart = 5)
}
