% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partGLS_methods.R
\name{part_ttest}
\alias{part_ttest}
\title{Correlated t-test for paritioned GLS}
\usage{
part_ttest(coefs, part.SEs, rcoef, df2, npart)
}
\arguments{
\item{coefs}{vector average GLS coefficients}

\item{part.SEs}{matrix of partition SEs for each coefficient (columns)}

\item{rcoef}{correlation among partition regression coefficients}

\item{df2}{second degree of freedom from partitioned GLS}

\item{npart}{number of partitions}
}
\value{
coefficient table with estimates, standard errors, t-statistics, and p-values
}
\description{
Correlated t-test for paritioned GLS
}
