% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{print.remoteCLS}
\alias{print.remoteCLS}
\title{printmethod for remoteCLS object}
\usage{
\method{print}{remoteCLS}(x, ...)
}
\arguments{
\item{x}{remoteCLS object}

\item{...}{additional arguments}
}
\description{
printmethod for remoteCLS object
}
