% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{print.remoteGLS}
\alias{print.remoteGLS}
\title{print method for remoteGLS}
\usage{
\method{print}{remoteGLS}(x, ret_call = TRUE, ...)
}
\arguments{
\item{x}{remoteGLS object}

\item{ret_call}{should the function call be returned?}

\item{...}{additional arguments}
}
\value{
formatted output for remoteGLS object
}
\description{
print method for remoteGLS
}
