% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteGLS_methods.R
\name{print.remoteGLS}
\alias{print.remoteGLS}
\title{print method for remoteGLS}
\usage{
\method{print}{remoteGLS}(x, digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{remoteGLS object}

\item{digits}{digits to print}

\item{...}{additional arguments}
}
\value{
formatted output for remoteGLS object
}
\description{
print method for remoteGLS
}
