% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remoteGLS_methods.R
\name{remoteGLS}
\alias{remoteGLS}
\title{remoteGLS constructor (S3)}
\usage{
remoteGLS(formula, formula0, no.F = FALSE)
}
\arguments{
\item{formula}{optional argument specifying the GLS formula}

\item{formula0}{optional argument specifying the null GLS formula}

\item{no.F}{optional argument specifying the no.F attribute}
}
\value{
an empty S3 object of class "remoteGLS"
}
\description{
remoteGLS constructor (S3)
}
\examples{
# tmp <- remoteGLS() #empty remoteGLS object

}
