% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotePARTS-package.R
\docType{package}
\name{remotePARTS-package}
\alias{remotePARTS}
\alias{remotePARTS-package}
\title{remotePARTS: Spatiotemporal Autoregression Analyses for Large Data Sets}
\description{
These tools were created to test map-scale hypotheses about trends in large remotely sensed data sets but any data with spatial and temporal variation can be analyzed. Tests are conducted using the PARTS method for analyzing spatially autocorrelated time series (Ives et al., 2021: \doi{10.1016/j.rse.2021.112678}). The method's unique approach can handle extremely large data sets that other spatiotemporal models cannot, while still appropriately accounting for spatial and temporal autocorrelation. This is done by partitioning the data into smaller chunks, analyzing chunks separately and then combining the separate analyses into a single, correlated test of the map-scale hypotheses.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/morrowcj/remotePARTS}
  \item Report bugs at \url{https://github.com/morrowcj/remotePARTS/issues}
}

}
\author{
\strong{Maintainer}: Clay Morrow \email{morrowcj@outlook.com} (\href{https://orcid.org/0000-0002-3069-3296}{ORCID})

Authors:
\itemize{
  \item Anthony Ives \email{arives@wisc.edu} (\href{https://orcid.org/0000-0001-9375-9523}{ORCID})
}

}
\keyword{internal}
