% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remotePARTS-package.R
\docType{package}
\name{remotePARTS-package}
\alias{remotePARTS}
\alias{remotePARTS-package}
\title{remotePARTS: Spatio-Temporal Auto-Regression for Remote Sensing Data}
\description{
"remotePARTS is a set of tools for analyzing sets of spatially correlated time series data, collected via remote sensing methods."
}
\author{
\strong{Maintainer}: Clay Morrow \email{morrow5@wisc.edu} (\href{https://orcid.org/0000-0002-3069-3296}{ORCID})

Authors:
\itemize{
  \item Anthony Ives \email{arives@wisc.edu} (\href{https://orcid.org/0000-0001-9375-9523}{ORCID})
}

}
\keyword{internal}
