% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{residuals.remoteAR}
\alias{residuals.remoteAR}
\title{Extract residuals from remoteAR object}
\usage{
\method{residuals}{remoteAR}(obj)
}
\arguments{
\item{obj}{remoteAR object}
}
\value{
residuals
}
\description{
Extract residuals from remoteAR object
}
