% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-functions.R
\name{sample_partitions}
\alias{sample_partitions}
\title{function to calculate partition size or number of partitions}
\usage{
sample_partitions(npix, npart = 10, partsize = NA, pixels = NA, verbose = TRUE)
}
\arguments{
\item{npix}{number of pixels in full dataset}

\item{pixels}{vector of pixel indexes to sample from}
}
\description{
function to calculate partition size or number of partitions
}
\examples{
# setup data
dat.M <- matrix(rnorm(3000*20), ncol = 20)
# 4 partitions (exhaustive)
sample_partitions(npix = nrow(dat.M), npart = 4)
# partitions with 500 pixels each (exhaustive)
sample_partitions(npix = nrow(dat.M), partsize = 500)
# 4 partitions each with 500 pixels (non-exhaustive)
sample_partitions(npix = nrow(dat.M), npart = 4, partsize = 500)

# index of pixels to subset
sub.indx <- 1:1000
# 4 partitions (exhaustive) using only the specified pixels
sample_partitions(npix = nrow(dat.M), npart = 4, pixels = sub.indx)
}
