% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/partitioned-GLS.R
\name{sample_partitions}
\alias{sample_partitions}
\title{Randomly sample a partition matrix for partitioned GLS}
\usage{
sample_partitions(npix, npart = 10, partsize = NA, pixels = NA, verbose = TRUE)
}
\arguments{
\item{npix}{number of pixels in full dataset}

\item{npart}{number of partitions to create}

\item{partsize}{size of each partition}

\item{pixels}{vector of pixel indexes to sample from}

\item{verbose}{logical: TRUE prints additional info}
}
\value{

}
\description{
Create a matrix whose columns contain indices of non-overlapping
random samples.
}
\details{

}
\examples{
# dummy data with 100 pixels and 20 time points
dat.M <- matrix(rnorm(100*20), ncol = 20)

# 4 partitions (exhaustive)
sample_partitions(npix = nrow(dat.M), npart = 4)

# partitions with 10 pixels each (exhaustive)
sample_partitions(npix = nrow(dat.M), partsize = 10)

# 4 partitions each with 10 pixels (non-exhaustive)
sample_partitions(npix = nrow(dat.M), npart = 4, partsize = 10)

# index of 50 pixels to subset
sub.indx <- c(1:10, 21:25, 30:62, 70:71)

# 5 partitions (exhaustive) using only the specified pixels
sample_partitions(npix = nrow(dat.M), npart = 5, pixels = sub.indx)

}
\seealso{
Other partitionedGLS: 
\code{\link{crosspart_GLS}()},
\code{\link{fitGLS_partition}()}
}
\concept{partitionedGLS}
