% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3-objects-and-methods.R
\name{summary.remoteCLS}
\alias{summary.remoteCLS}
\title{summary of remoteCLS object}
\usage{
\method{summary}{remoteCLS}(object, ...)
}
\arguments{
\item{object}{remoteCLS object}

\item{...}{additional arguments}
}
\description{
summary of remoteCLS object
}
