% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLS.R
\name{V.fit}
\alias{V.fit}
\title{Fit varcov matrix from distance matrix}
\usage{
V.fit(Dist, spatialcor, FUN = "exponential")
}
\arguments{
\item{Dist}{n x n numeric distance matrix}

\item{spatialcor}{spatial correlation parameter(s)}

\item{fun}{function with which to transform the distance matrix into the
varcov matrix}
}
\value{

}
\description{
Fit varcov matrix from distance matrix
}
\details{
Note: currently the "exponential" version yields a non-positive definitive
(and therefore invalid) V matrix.

Additionally, V.fit() is functionally identical to fitV() however, fitV()
should be used instead. This function should be phased out
}
\examples{
#TBA
}
